requirements.txt
pyserial>=3.5
pymeasure>=0.13
